adds offset class and tests
